FIX users controllers and FEAT utils of resultUsers and registerUsers

+++ Backend/src/api/controllers/users.js
@@ -1,22 +1,26 @@
+const bcrypt = require("bcrypt");
+const { registertUserControlError } = require("../../utils/registerUserControlError");
 const User = require("../models/users");
+const { generateToken } = require("../../utils/jwt");
+const { resultAllUsers } = require("../../utils/resultAllUsers");
+const { resultUsersByName } = require("../../utils/resultUsersByName");
+const { resultUsersByPhone } = require("../../utils/resultUsersByPhone");
+const { resultUserDeleted } = require("../../utils/resultUserDeleted");
 
 const registerUser = async (req, res, next) => {
     try {
         const { name, email, password, phone } = req.body;
-
-        if (name.length < 2 || name.length > 20) {
-            return res.status(400).json({ message: "La nombre debe de tener de 2 a 20 caracteres." });
-        }
-        if (password.length < 8 || password.length > 16) {
-            return res.status(400).json({ message: "La password debe de entre 8 y 16 caracteres." });
-        }
-        if (phone.length !== 9) {
-            return res.status(400).json({ message: "El número de teléfono debe de tener 9 digitos." });
-        }
+        registertUserControlError(res, name, password, phone, email);
 
         const userDuplicated = await User.findOne({ $or: [{ name }, { email }, { phone }] });
         if (userDuplicated) {
-            return res.status(400).json({ message: "El nombre o el email o el teléfono ya están en uso por otro usuario." });
+            if (userDuplicated.name === name) {
+                return res.status(400).json({ message: "El nombre ya están en uso por otro usuario." });
+            } else if (userDuplicated.email == email) {
+                return res.status(400).json({ message: "El email ya está registrado por otro usuario." });
+            } else if (userDuplicated.phone == phone) {
+                return res.status(400).json({ message: "El número de teléfono ya está registrado por otro usuario." });
+            }
         }
 
         const newUser = new User(req.body);
@@ -27,10 +31,36 @@ const registerUser = async (req, res, next) => {
     }
 };
 
+const loginUser = async (req, res, next) => {
+    try {
+        const { email, password } = req.body;
+        const userLogin = await User.findOne({ email });
+        if (userLogin) {
+            if (bcrypt.compareSync(password, userLogin.password)) {
+                const token = generateToken(userLogin._id);
+                return res.status(200).json({ message: "LOGIN realizado correctamente.", userLogin, token });
+            } else {
+                return res.status(400).json({ message: "El email o la contraseña son incorrectos." });
+            }
+        } else {
+            return res.status(400).json({ message: "El email no existe." });
+        }
+    } catch (error) {
+        return res.status(400).json(`❌ Fallo en loginUser: ${error.message}`);
+    }
+};
+
 const getAllUsers = async (req, res, next) => {
     try {
-        const allUsers = await User.find().select("-password, -role, -email").populate("match");
-        allUsers.length ? res.status(200).json({ message: "Listado completo de usuarios.", allUsers }) : res.status(400).json({ message: "¡Hola! Eres el único usuario que existe... por ahora..." });
+        const user = req.user;
+
+        if (user.role === "admin") {
+            const allUsers = await User.find().populate("match");
+            resultAllUsers(res, allUsers);
+        } else {
+            const allUsers = await User.find().select("-password -role -email").populate("match");
+            resultAllUsers(res, allUsers);
+        }
     } catch (error) {
         return res.status(400).json(`❌ Fallo en getAllUsers: ${error.message}`);
     }
@@ -39,10 +69,17 @@ const getAllUsers = async (req, res, next) => {
 const getUserByName = async (req, res, next) => {
     try {
         const { name } = req.params;
-        const searchUserByName = await User.find({ name: new RegExp(name, "i") }).select("-password -role -email").populate("match");
-        searchUserByName.length
-            ? res.status(200).json({ message: `Listado de usuarios con el nombre ${name} : `, searchUserByName })
-            : res.status(400).json({ message: `No se ha encontrado ningún usuario con ese nombre: ${name}` });
+        const user = req.user;
+
+        if (user.role === "admin") {
+            const searchUserByName = await User.find({ name: new RegExp(name, "i") }).populate("match");
+            resultUsersByName(res, searchUserByName, name);
+        } else {
+            const searchUserByName = await User.find({ name: new RegExp(name, "i") })
+                .select("-password -role -email")
+                .populate("match");
+            resultUsersByName(res, searchUserByName, name);
+        }
     } catch (error) {
         return res.status(400).json(`❌ Fallo en getUserByName: ${error.message}`);
     }
@@ -51,13 +88,18 @@ const getUserByName = async (req, res, next) => {
 const getUserByPhone = async (req, res, next) => {
     try {
         const { phone } = req.params;
+        const user = req.user;
+
         if (phone.length !== 9) {
             return res.status(400).json("Introduce un número de teléfono de 9 digitos.");
         }
-        const searchUserByPhone = await User.find({ phone }).select("-password, -role, -email").populate("match");
-        searchUserByPhone.length
-            ? res.status(200).json({ message: `El número de teléfono ${phone} corresponde al usuario: `, searchUserByPhone })
-            : res.status(400).json({ message: `No se ha encontrado ningún usuario con el número de teléfono ${phone}` });
+        if (user.role === "admin") {
+            const searchUserByPhone = await User.find({ phone }).populate("match");
+            resultUsersByPhone(res, searchUserByPhone, phone);
+        } else {
+            const searchUserByPhone = await User.find({ phone }).select("-password -role -email").populate("match");
+            resultUsersByPhone(res, searchUserByPhone, phone);
+        }
     } catch (error) {
         return res.status(400).json(`❌ Fallo en getUserByPhone: ${error.message}`);
     }
@@ -65,8 +107,7 @@ const getUserByPhone = async (req, res, next) => {
 
 const updateUser = async (req, res, next) => {
     try {
-        const {  } = req.body;
-
+        const {} = req.body;
     } catch (error) {
         return res.status(400).json(`❌ Fallo en updateUser: ${error.message}`);
     }
@@ -75,12 +116,16 @@ const updateUser = async (req, res, next) => {
 const deleteUser = async (req, res, next) => {
     try {
         const { id } = req.params;
-        const userDeleted = await User.findByIdAndDelete(id)
-        userDeleted ? res.status(200).json({ message: `Usuario eliminado correctamente.`, userDeleted })
-        : res.status(400).json({ message: `No se puede eliminar ese usuario ya que no existe.` });
+        const user = req.user
+
+        if (user._id.toString() !== id && user.role !== 'admin') {
+            return res.status(400).json({ message: 'No tienes permisos de Administrador para eliminar el usuario.' })
+        }
+        const userDeleted = await User.findByIdAndDelete(id);
+        resultUserDeleted(res, userDeleted)
     } catch (error) {
         return res.status(400).json(`❌ Fallo en deleteUser: ${error.message}`);
     }
 };
 
-module.exports = { registerUser, getAllUsers, getUserByName, getUserByPhone, updateUser, deleteUser };
+module.exports = { registerUser, loginUser, getAllUsers, getUserByName, getUserByPhone, updateUser, deleteUser };
